add LogDataSourceErrorAspect and @LogDataSourceError

+++ src/main/java/ru/t1/java/demo/dto/TransactionDto.java
@@ -2,12 +2,14 @@ package ru.t1.java.demo.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.persistence.Column;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -24,7 +26,7 @@ public class TransactionDto implements Serializable {
     @JsonProperty("account_id")
     private Long account_id;
     @JsonProperty("amount")
-    private Double amount;
+    private BigDecimal amount;
     @JsonProperty("transactionTime")
     private LocalDateTime transactionTime;
 }
